Restructuring project_ui vs project
