+++ data/3aaf92677e614f58810378233deba1d2.json
@@ -1,0 +1,1 @@
+{"translated_word": "eccdfcaaff", "translated_word_components": "", "in_language_components": "", "etymological_symbology": "", "compiled_symbology": "", "symbol_mapping": "", "symbol_selection": "", "symbol_pattern_selected": "", "rules_applied": "", "in_language_word": "", "version_history": "", "has_been_modified_since_last_resolve": "", "has_modified_ancestor": ""}
