+++ src/core/project.py
@@ -1,0 +1,3 @@
+"""Library for Project level functionality"""
+# A project HAS a Settings Instance
+# A project HAS indexed Lexicon Instances
